Split mapper.sharedcode -> mapper.sharedcode.codes, mapper.sharedcode.variables
