+++ internal/pkg/mapper/sharedcode/codes/load.go
@@ -1,4 +1,4 @@
-package sharedcode
+package codes
 
 import (
 	"github.com/keboola/keboola-as-code/internal/pkg/model"
@@ -13,48 +13,8 @@ type loader struct {
 	errors    *utils.Error
 }
 
-// MapAfterRemoteLoad - extract shared code "variables_id".
-func (m *sharedCodeMapper) MapAfterRemoteLoad(recipe *model.RemoteLoadRecipe) error {
-	// Variables are used by shared code - config row.
-	apiObject, ok := recipe.ApiObject.(*model.ConfigRow)
-	if !ok {
-		return nil
-	}
-	internalObject := recipe.InternalObject.(*model.ConfigRow)
-
-	// Check component type
-	component, err := m.State.Components().Get(internalObject.ComponentKey())
-	if err != nil {
-		return err
-	}
-	if !component.IsSharedCode() {
-		return nil
-	}
-
-	// Variables ID is stored in configuration
-	variablesIdRaw, found := apiObject.Content.Get(model.SharedCodeVariablesIdContentKey)
-	if !found {
-		return nil
-	}
-
-	// Variables ID must be string
-	variablesId, ok := variablesIdRaw.(string)
-	if !ok {
-		return nil
-	}
-
-	// Create relation
-	internalObject.AddRelation(&model.SharedCodeVariablesFromRelation{
-		VariablesId: variablesId,
-	})
-
-	// Remove variables ID from configuration content
-	internalObject.Content.Delete(model.SharedCodeVariablesIdContentKey)
-	return nil
-}
-
 // MapAfterLocalLoad - load shared code from filesystem to target config.
-func (m *sharedCodeMapper) MapAfterLocalLoad(recipe *model.LocalLoadRecipe) error {
+func (m *mapper) MapAfterLocalLoad(recipe *model.LocalLoadRecipe) error {
 	// Only for shared code config row
 	if ok, err := m.isSharedCodeConfigRow(recipe.Object); err != nil {
 		return err
