+++ internal/pkg/mapper/sharedcode/codes/save.go
@@ -1,4 +1,4 @@
-package sharedcode
+package codes
 
 import (
 	"fmt"
@@ -16,43 +16,7 @@ type writer struct {
 	errors    *utils.Error
 }
 
-// MapBeforeRemoteSave - add "variables_id" to shared code.
-func (m *sharedCodeMapper) MapBeforeRemoteSave(recipe *model.RemoteSaveRecipe) error {
-	// Variables are used by shared code - config row.
-	internalObject, ok := recipe.InternalObject.(*model.ConfigRow)
-	if !ok {
-		return nil
-	}
-	apiObject := recipe.ApiObject.(*model.ConfigRow)
-
-	// Check component type
-	component, err := m.State.Components().Get(internalObject.ComponentKey())
-	if err != nil {
-		return err
-	}
-	if !component.IsSharedCode() {
-		return nil
-	}
-
-	// Get relation
-	relType := model.SharedCodeVariablesFromRelType
-	relationRaw, err := internalObject.Relations.GetOneByType(relType)
-	if err != nil {
-		return fmt.Errorf(`unexpected state of %s: %w`, recipe.Manifest.Desc(), err)
-	} else if relationRaw == nil {
-		return nil
-	}
-	relation := relationRaw.(*model.SharedCodeVariablesFromRelation)
-
-	// Set variables ID
-	apiObject.Content.Set(model.SharedCodeVariablesIdContentKey, relation.VariablesId)
-
-	// Delete relation
-	apiObject.Relations.RemoveByType(relType)
-	return nil
-}
-
-func (m *sharedCodeMapper) MapBeforeLocalSave(recipe *model.LocalSaveRecipe) error {
+func (m *mapper) MapBeforeLocalSave(recipe *model.LocalSaveRecipe) error {
 	// Only for shared code config row
 	if ok, err := m.isSharedCodeConfigRow(recipe.Object); err != nil {
 		return err
