+++ internal/pkg/mapper/sharedcode/variables/variables_test.go
@@ -1,0 +1,22 @@
+package variables_test
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"go.uber.org/zap"
+
+	"github.com/keboola/keboola-as-code/internal/pkg/filesystem/aferofs"
+	"github.com/keboola/keboola-as-code/internal/pkg/model"
+	"github.com/keboola/keboola-as-code/internal/pkg/testapi"
+	"github.com/keboola/keboola-as-code/internal/pkg/utils"
+)
+
+func createMapperContext(t *testing.T) model.MapperContext {
+	t.Helper()
+	logger, _ := utils.NewDebugLogger()
+	fs, err := aferofs.NewMemoryFs(logger, ".")
+	assert.NoError(t, err)
+	state := model.NewState(zap.NewNop().Sugar(), fs, model.NewComponentsMap(testapi.NewMockedComponentsProvider()), model.SortByPath)
+	return model.MapperContext{Logger: logger, Fs: fs, Naming: model.DefaultNaming(), State: state}
+}
